Tunnel user change reflected
-So I changed it so that there isn't a place where a User is stored.
Instead everything being stored is being stored as a Tunnel. Then I made
the changes so that that is reflected everywhere that users was directly
called. This is not a cause of worry, cause when we are accessing the
information from the db we use a dbset created in datacontext. This
narrows the pool of searches even if they are all tunnels. If and when
the firebase equivalent is to be made the firebase context has to
utilize the same kind of logic (dbset )
-I also installed for the first time a couple of extensions, including
TODO highlighter. And learnt about shortcuts and Task List