Changed TunnelId type and add a note file
-So I was having trouble working with the idGenerator I created so I
decided to change TunnelId to string so that it fits. I didn't want to
use slong or something just to store a number so string is the best bet.
-Following the reflective changes need to be done cause of TunnelId I
wanted to make notes as I progressed over the thoughts I was having with
the whole project in its entirety. I didnt want to put it in the README
so I made my own text file called note to myself.
-I also added a Task list token called Note so where-ever notes are
being made it will be captured by the Task List