Removed AbstractTunnel and removed the Typeparams
-So I learnt the EF doesn't really know what to do with types with type
parameters. So I had to change my code so that it Tunnel acts like
AbstractTunnel.
-I also changed some things in the Tunnel constructor so that we can set
the TunnelId. We don't need to depend on the switch. We will just
override a method that Tunnels will inhert. So don't forget to set that
there and no where else.
-Changes were reflected in the controllers based on the previously
mentioned change
-Every Single class that inherits from Tunnel is expected to have its
constructor derive from the base class so that the TunnelId will be set.
-