Added Avatar in the database and made a recordi
-So I managed to add Avatar. The error came cause I don't think EF Core
knows how to deal with collections with primitive types. So I made a
record that has the string property that I wanted to make a list of and
then made a collection of that in Avatar.
-I also changed the access modifier of MakeNewTunnel in Tunnel cause I
wanted methods deep below the immediate subclass to access it, but not
all assemblies. It is prone to revise but thats fine